Working on templates

+++ src/main/java/com/github/percivalgebashe/assignment_5/repository/BookRepository.java
@@ -4,5 +4,5 @@ import com.github.percivalgebashe.assignment_5.entity.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
 
-public interface BookRepository extends JpaRepository<Book, Integer>, CustomBookRepository, JpaSpecificationExecutor<Book> {
+public interface BookRepository extends JpaRepository<Book, Long>, CustomBookRepository, JpaSpecificationExecutor<Book> {
 }
